# services/core-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Vendor:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    High Performance Backend Service for PT. Adijayantara Logistics Indonesia.
    Architecture & Headers:
    - X-API-Version: Target API versioning (e.g. 'v1', 'v2')
    - X-Idempotency-Key: Unique UUID token for safe mutations (POST/PUT/PATCH/DELETE) to prevent duplicates
    - X-Request-ID: Distributed tracing correlation ID
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cashflow & Shipment Management API
  version: "1.0"
paths:
...
  /vendors:
    get:
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all vendors
      tags:
      - vendors
    post:
      consumes:
      - application/json
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Vendor Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Vendor'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create new vendor
      tags:
      - vendors
  /vendors/{id}:
    delete:
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Vendor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft delete vendor by ID
      tags:
      - vendors
    get:
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Vendor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get vendor by ID
      tags:
      - vendors
    put:
      consumes:
      - application/json
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Vendor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Vendor Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Vendor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update vendor by ID
      tags:
      - vendors
securityDefinitions:
  ApiVersion:
    description: Target API Version Routing (e.g. v1, v2)
    in: header
    name: X-API-Version
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token.
    in: header
    name: Authorization
    type: apiKey
  IdempotencyKey:
    description: Unique UUID idempotency key for mutation operations (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Idempotency-Key
    type: apiKey
swagger: "2.0"